+++ cmd/retro/runner.go
@@ -24,6 +24,8 @@ func (r Runner) preflight() (copyHTML func(string, string, string) error, err er
 
 	// Set env vars
 	switch cmd {
+
+	// TODO: Add support for source map, etc.
 	case KindDevCommand:
 		os.Setenv("CMD", "dev")
 		os.Setenv("ENV", "development")
@@ -32,6 +34,7 @@ func (r Runner) preflight() (copyHTML func(string, string, string) error, err er
 		os.Setenv("SRC_DIR", SRC_DIR)
 		os.Setenv("OUT_DIR", OUT_DIR)
 
+	// TODO: Add support for source map, etc.
 	case KindBuildCommand:
 		os.Setenv("CMD", "build")
 		os.Setenv("ENV", "production")
@@ -40,6 +43,7 @@ func (r Runner) preflight() (copyHTML func(string, string, string) error, err er
 		os.Setenv("SRC_DIR", SRC_DIR)
 		os.Setenv("OUT_DIR", OUT_DIR)
 
+	// TODO: Add support for source map, etc.
 	case KindServeCommand:
 		os.Setenv("CMD", "serve")
 		os.Setenv("ENV", "production")
@@ -62,8 +66,8 @@ func (r Runner) preflight() (copyHTML func(string, string, string) error, err er
 	if err := cpdir(WWW_DIR, filepath.Join(OUT_DIR, WWW_DIR), []string{filepath.Join(WWW_DIR, "index.html")}); err != nil {
 		return nil, err
 	}
-	copyHTML = copyHTMLEntryPoint
 
+	copyHTML = copyHTMLEntryPoint
 	return
 }